+++ li-cli/li_cli/commands/__init__.py
@@ -1,1 +1,0 @@
-"""li-cli commands."""